- Move Senso discovery code to own module - Make replay tool discoverable - Add tool for simply discovering Sensos
