+++ src-test/discover.js
@@ -1,0 +1,3 @@
+// Tool to discover Sensos
+
+require('../src/Senso/discovery')()
